% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalTuningParam.R
\name{plot.evalTuningParam}
\alias{plot.evalTuningParam}
\title{Plots the tuning surface of TPLS}
\usage{
\method{plot}{evalTuningParam}(object)
}
\arguments{
\item{object}{: evalTuningParam object}
}
\description{
Plots the tuning surface of TPLS
}
\examples{
# See examples for TPLS_cv
}
