% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalTuningParam.R
\name{plotTuningSurface}
\alias{plotTuningSurface}
\title{Plots the tuning surface of TPLS}
\usage{
plotTuningSurface(object)
}
\arguments{
\item{object}{: evalTuningParam object}
}
\description{
Plots the tuning surface of TPLS
}
