% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPLS.R
\name{predict.TPLS}
\alias{predict.TPLS}
\title{Make predictions about given data \code{testX} by using an extracted TPLSmdl with \code{compval} components and \code{threshval} threshold.}
\usage{
\method{predict}{TPLS}(TPLSmdl, compval, threshval, testX)
}
\arguments{
\item{TPLSmdl}{A TPLS object created from using function \code{TPLS}}

\item{compval}{The number of components you want in your model. Providing a vector will provide multiple predictions (e.g., c(3,4,5) will provide three prediction columns each with 3, 4, and 5 PLS components)}

\item{threshval}{Threshold number between 0 and 1 (inclusive) for thresholding the betamap. This must be a scalar.}
}
\value{
\itemize{
    \item \code{score}: Column vector of prediction scores. Matrix of scores if compval is a vector.
}
}
\description{
Make predictions about given data \code{testX} by using an extracted TPLSmdl with \code{compval} components and \code{threshval} threshold.
}
\examples{
# See examples for TPLS
}
