% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPLS_cv.R
\name{TPLS_cv}
\alias{TPLS_cv}
\title{Fit a TPLS model to data with cross validation}
\usage{
TPLS_cv(X, Y, foldid, NComp = 50, W = 0)
}
\arguments{
\item{X}{n-by-v data matrix of real numbers. Rows correspond to observations (trials) and columns to variables (e.g., fMRI voxels).}

\item{Y}{n-by-1 Vector of real numbers. Can be binary (0/1) for classification model, or can be continuous.}

\item{foldid}{A vector of values between 1 and number of folds identifying what fold each observation is in.}

\item{NComp}{Maximum number of partial least squares component you want to use. Default is 50, and this is on the safe side for fMRI.}

\item{W}{n-by-1 vector of positive observation weights.}
}
\value{
A TPLS_cv object that contains the following attributes. Most of the time, you won't need to access the attributes.
\itemize{
    \item \code{NComp}: The number of components you specified in the input
    \item \code{numfold}: Total number of cross-validation folds
    \item \code{testfold}: A vector indice that should be the same as foldid, if it was provided accurately.
    \item \code{cvMdls} : A vector of TPLS models, one for each fold.
}
}
\description{
Fit a TPLS model to data with cross validation
}
\examples{
# Fit TPLS model to example data
}
