% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPLS.R
\name{TPLSpredict}
\alias{TPLSpredict}
\title{Method for making predictions on a testing dataset testX}
\usage{
TPLSpredict(TPLSmdl, compval, threshval, testX)
}
\arguments{
\item{TPLSmdl}{A TPLS object created from using function \code{TPLS}}

\item{compval}{Vector of number of components to use in final predictor. Providing a vector will provide multiple predictions (e.g., c(3,4,5) will provide three prediction columns each with 3, 4, and 5 PLS components)}

\item{threshval}{Threshold number between 0 and 1 (inclusive) for thresholding the betamap. This must be a scalar.}

\item{testX}{Data that you want to predict the Y of}
}
\value{
\itemize{
    \item \code{score}: Column vector of prediction scores. Matrix of scores if compval is a vector.
}
}
\description{
Method for making predictions on a testing dataset testX
}
